[fix] Refactoring IO.cs
