+++ src/EasyEPlanner.csproj
@@ -195,8 +195,12 @@
       <DependentUpon>IdleForm.cs</DependentUpon>
     </Compile>
     <Compile Include="IdleModule.cs" />
-    <Compile Include="IO.cs" />
     <Compile Include="IOReader.cs" />
+    <Compile Include="IO\IOLinkCalculator.cs" />
+    <Compile Include="IO\IOManager.cs" />
+    <Compile Include="IO\IOModule.cs" />
+    <Compile Include="IO\IOModuleInfo.cs" />
+    <Compile Include="IO\IONode.cs" />
     <Compile Include="Logs\LogFrm.cs">
       <SubType>Form</SubType>
     </Compile>
